@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -23,23 +23,30 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
+
+  # state management
   stacked:
     path: ../
-
-  # dependency injection
-  get_it: ^4.0.1
-  injectable: ^0.3.0 
+  stacked_services:
+    path: ../../stacked_services/
+  stacked_hooks:
+  flutter_hooks:
+  stacked_themes:
 
   # navigation
   get:
+
+  # logging
+  logger:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   # dependency injection
-  injectable_generator:
   build_runner:
+  stacked_generator: #^0.2.6
+    path: ../../stacked_generator
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
