@@ -1,16 +1,16 @@
 name: stacked_hooks
 description: Classes that help you make use of hooks when using the stacked package
-version: 0.1.3+2
+version: 0.1.4
 homepage: https://github.com/FilledStacks/stacked/tree/master/packages/stacked_hooks
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0-133.7.beta <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.15.0
-  provider: ^4.3.1
+  flutter_hooks: ^0.16.0
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
