@@ -1,26 +1,26 @@
 name: stacked_themes
 description: A set of classes to help you better manage Themes in flutter
-version: 0.2.2+2
+version: 0.2.4
 homepage: https://github.com/FilledStacks/stacked/tree/master/packages/stacked_themes
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.3.2+3
-  rxdart: ^0.25.0
-  shared_preferences: ^0.5.12+4
-  get_it: ^5.0.3
+  provider: ^5.0.0
+  rxdart: ^0.26.0
+  shared_preferences: ^2.0.4
+  get_it: ^6.0.0
   flutter_statusbarcolor: ^0.2.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  mockito:
+  mockito: ^5.0.0-nullsafety.7
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
