@@ -1,6 +1,6 @@
 class Human {
-  final String name;
-  final String surname;
+  final String? name;
+  final String? surname;
 
   Human({this.name, this.surname});
 }
