@@ -7,7 +7,7 @@ import 'favorites/favorites_view.dart';
 import 'history/history_view.dart';
 
 class BottomNavExample extends StatefulWidget {
-  const BottomNavExample({Key key}) : super(key: key);
+  const BottomNavExample({Key? key}) : super(key: key);
 
   @override
   _BottomNavExampleState createState() => _BottomNavExampleState();
@@ -61,6 +61,6 @@ class _BottomNavExampleState extends State<BottomNavExample> {
       }
     }
 
-    return _viewCache[index];
+    return _viewCache[index]!;
   }
 }
