@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:new_architecture/app/locator.dart';
+import 'package:new_architecture/app/app.locator.dart';
 import 'package:stacked/stacked.dart';
 
 import 'favorites_viewmodel.dart';
 
 class FavoritesView extends StatelessWidget {
-  const FavoritesView({Key key}) : super(key: key);
+  const FavoritesView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
