@@ -4,14 +4,13 @@ import 'package:stacked/stacked.dart';
 import 'profile_viewmodel.dart';
 
 class ProfileView extends StatelessWidget {
-  const ProfileView({Key key}) : super(key: key);
+  const ProfileView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder<ProfileViewModel>.reactive(
-      builder: (context, model, child) => Scaffold(
-        body: Center(child: Text('History'))
-      ),
+      builder: (context, model, child) =>
+          Scaffold(body: Center(child: Text('History'))),
       viewModelBuilder: () => ProfileViewModel(),
     );
   }
