@@ -3,7 +3,7 @@ import 'package:new_architecture/ui/startup/startup_viewmodel.dart';
 import 'package:stacked/stacked.dart';
 
 class StartupView extends StatelessWidget {
-  const StartupView({Key key}) : super(key: key);
+  const StartupView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
