@@ -10,7 +10,7 @@ import 'bottom_sheet_ui.dart';
 /// A service that allows you to show a bottom sheet
 class BottomSheetService {
   Map<dynamic,
-          Widget Function(BuildContext, SheetRequest, Function(SheetResponse))>
+          Widget Function(BuildContext, SheetRequest, Function(SheetResponse))>?
       _sheetBuilders;
 
   void setCustomSheetBuilders(
@@ -22,13 +22,13 @@ class BottomSheetService {
     _sheetBuilders = builders;
   }
 
-  Future<SheetResponse> showBottomSheet({
-    @required String title,
-    String description,
+  Future<SheetResponse?> showBottomSheet({
+    required String title,
+    String? description,
     String confirmButtonTitle = 'Ok',
-    String cancelButtonTitle,
+    String? cancelButtonTitle,
   }) {
-    return Get.bottomSheet<SheetResponse>(
+    return Get.bottomSheet<SheetResponse?>(
       Material(
         type: MaterialType.transparency,
         child: GeneralBottomSheet(
@@ -40,7 +40,7 @@ class BottomSheetService {
           onCancelTapped: () => completeSheet(SheetResponse(confirmed: false)),
         ),
       ),
-      backgroundColor: Theme.of(Get.context).brightness == Brightness.light
+      backgroundColor: Theme.of(Get.context!).brightness == Brightness.light
           ? Colors.white
           : Colors.grey[800],
       shape: RoundedRectangleBorder(
@@ -53,60 +53,65 @@ class BottomSheetService {
   }
 
   // Creates a popup with the given widget, a scale animation, and faded background.
-  Future<SheetResponse> showCustomSheet({
+  Future<SheetResponse?> showCustomSheet({
     dynamic variant,
-    String title,
-    String description,
+    String? title,
+    String? description,
     bool hasImage = false,
-    String imageUrl,
+    String? imageUrl,
     bool showIconInMainButton = false,
-    String mainButtonTitle,
+    String? mainButtonTitle,
     bool showIconInSecondaryButton = false,
-    String secondaryButtonTitle,
+    String? secondaryButtonTitle,
     bool showIconInAdditionalButton = false,
-    String additionalButtonTitle,
+    String? additionalButtonTitle,
     bool takesInput = false,
     Color barrierColor = Colors.black54,
     bool barrierDismissible = false,
+    bool isScrollControlled = false,
     String barrierLabel = '',
     dynamic customData,
   }) {
-    final sheetBuilder = _sheetBuilders[variant];
-
     assert(
-      sheetBuilder != null,
+      _sheetBuilders != null,
       '''
       There's no sheet builder supplied for the variant:$variant. If you haven't yet setup your
       custom builder. Please call the setCustomSheetBuilders function on the service and supply
-      the UI that you'd like to build for each variant. 
+      the UI that you'd like to build for each variant.
 
       If you have already done that. Make sure that the variant:$variant has a builder associated
       with it.
       ''',
     );
 
-    return Get.bottomSheet<SheetResponse>(Material(
-      type: MaterialType.transparency,
-      child: sheetBuilder(
-        Get.context,
-        SheetRequest(
-          title: title,
-          description: description,
-          hasImage: hasImage,
-          imageUrl: imageUrl,
-          showIconInMainButton: showIconInMainButton,
-          mainButtonTitle: mainButtonTitle,
-          showIconInSecondaryButton: showIconInSecondaryButton,
-          secondaryButtonTitle: secondaryButtonTitle,
-          showIconInAdditionalButton: showIconInAdditionalButton,
-          additionalButtonTitle: additionalButtonTitle,
-          takesInput: takesInput,
-          customData: customData,
-          variant: variant,
+    final sheetBuilder = _sheetBuilders![variant];
+
+    return Get.bottomSheet<SheetResponse>(
+      Material(
+        type: MaterialType.transparency,
+        child: sheetBuilder!(
+          Get.context!,
+          SheetRequest(
+            title: title,
+            description: description,
+            hasImage: hasImage,
+            imageUrl: imageUrl,
+            showIconInMainButton: showIconInMainButton,
+            mainButtonTitle: mainButtonTitle,
+            showIconInSecondaryButton: showIconInSecondaryButton,
+            secondaryButtonTitle: secondaryButtonTitle,
+            showIconInAdditionalButton: showIconInAdditionalButton,
+            additionalButtonTitle: additionalButtonTitle,
+            takesInput: takesInput,
+            customData: customData,
+            variant: variant,
+          ),
+          completeSheet,
         ),
-        completeSheet,
       ),
-    ));
+      isDismissible: barrierDismissible,
+      isScrollControlled: isScrollControlled,
+    );
   }
 
   /// Completes the dialog and passes the [response] to the caller
