@@ -18,7 +18,7 @@ class DialogService {
   Map<
           dynamic,
           Widget Function(
-              BuildContext, DialogRequest, Function(DialogResponse))>
+              BuildContext, DialogRequest, Function(DialogResponse))>?
       _dialogBuilders;
 
   void registerCustomDialogBuilders(
@@ -53,10 +53,10 @@ class DialogService {
     'Prefer to use the registerCustomDialogBuilders() method. This method will be removed on the next major release. 0.7.0',
   )
   void registerCustomDialogBuilder({
-    @required dynamic variant,
-    @required
-        Widget Function(BuildContext, DialogRequest, Function(DialogResponse))
-            builder,
+    required dynamic variant,
+    required Widget Function(
+            BuildContext, DialogRequest, Function(DialogResponse))
+        builder,
   }) {
     _customDialogBuilders[variant] = builder;
   }
@@ -64,17 +64,19 @@ class DialogService {
   /// Shows a dialog to the user
   ///
   /// It will show a platform specific dialog by default. This can be changed by setting [dialogPlatform]
-  Future<DialogResponse> showDialog({
-    String title,
-    String description,
-    String cancelTitle,
+  Future<DialogResponse?> showDialog({
+    String? title,
+    String? description,
+    String? cancelTitle,
+    Color? cancelTitleColor,
     String buttonTitle = 'Ok',
+    Color? buttonTitleColor,
     bool barrierDismissible = false,
 
     /// Indicates which [DialogPlatform] to show.
     ///
     /// When not set a Platform specific dialog will be shown
-    DialogPlatform dialogPlatform,
+    DialogPlatform? dialogPlatform,
   }) {
     if (dialogPlatform != null) {
       return _showDialog(
@@ -93,19 +95,23 @@ class DialogService {
         title: title,
         description: description,
         cancelTitle: cancelTitle,
+        cancelTitleColor: cancelTitleColor,
         buttonTitle: buttonTitle,
+        buttonTitleColor: buttonTitleColor,
         dialogPlatform: _dialogType,
         barrierDismissible: barrierDismissible,
       );
     }
   }
 
-  Future<DialogResponse> _showDialog({
-    String title,
-    String description,
-    String cancelTitle,
-    String buttonTitle,
-    DialogPlatform dialogPlatform,
+  Future<DialogResponse?> _showDialog({
+    String? title,
+    String? description,
+    String? cancelTitle,
+    Color? cancelTitleColor,
+    String? buttonTitle,
+    Color? buttonTitleColor,
+    DialogPlatform dialogPlatform = DialogPlatform.Material,
     bool barrierDismissible = false,
   }) {
     var isConfirmationDialog = cancelTitle != null;
@@ -118,7 +124,8 @@ class DialogService {
           if (isConfirmationDialog)
             PlatformButton(
               dialogPlatform: dialogPlatform,
-              text: cancelTitle,
+              text: cancelTitle!,
+              cancelBtnColor: cancelTitleColor,
               isCancelButton: true,
               onPressed: () {
                 completeDialog(
@@ -130,7 +137,8 @@ class DialogService {
             ),
           PlatformButton(
             dialogPlatform: dialogPlatform,
-            text: buttonTitle,
+            text: buttonTitle!,
+            confirmationBtnColor: buttonTitleColor,
             onPressed: () {
               completeDialog(
                 DialogResponse(
@@ -146,27 +154,30 @@ class DialogService {
   }
 
   // Creates a popup with the given widget, a scale animation, and faded background.
-  Future<DialogResponse> showCustomDialog({
+  Future<DialogResponse?> showCustomDialog({
     dynamic variant,
-    String title,
-    String description,
+    String? title,
+    String? description,
     bool hasImage = false,
-    String imageUrl,
+    String? imageUrl,
     bool showIconInMainButton = false,
-    String mainButtonTitle,
+    String? mainButtonTitle,
     bool showIconInSecondaryButton = false,
-    String secondaryButtonTitle,
+    String? secondaryButtonTitle,
     bool showIconInAdditionalButton = false,
-    String additionalButtonTitle,
+    String? additionalButtonTitle,
     bool takesInput = false,
     Color barrierColor = Colors.black54,
     bool barrierDismissible = false,
     String barrierLabel = '',
     dynamic customData,
   }) {
-    // TODO: Remove the _customDialogUIBuilders in the next major release 0.7.0
-    final customDialogUI =
-        _dialogBuilders[variant] ?? _customDialogBuilders[variant];
+    assert(
+      _dialogBuilders != null,
+      'You have to call registerCustomDialogBuilder to use this function. Look at the custom dialog UI section in the stacked_services readme.',
+    );
+
+    final customDialogUI = _dialogBuilders![variant];
 
     assert(
       customDialogUI != null,
@@ -181,7 +192,7 @@ class DialogService {
       useRootNavigator: true,
       pageBuilder: (BuildContext buildContext, _, __) => SafeArea(
         child: Builder(
-          builder: (BuildContext context) => customDialogUI(
+          builder: (BuildContext context) => customDialogUI!(
             context,
             DialogRequest(
               title: title,
@@ -216,9 +227,9 @@ class DialogService {
   }
 
   /// Shows a confirmation dialog with title and description
-  Future<DialogResponse> showConfirmationDialog({
-    String title,
-    String description,
+  Future<DialogResponse?> showConfirmationDialog({
+    String? title,
+    String? description,
     String cancelTitle = 'Cancel',
     String confirmationTitle = 'Ok',
     bool barrierDismissible = false,
@@ -226,7 +237,7 @@ class DialogService {
     /// Indicates which [DialogPlatform] to show.
     ///
     /// When not set a Platform specific dialog will be shown
-    DialogPlatform dialogPlatform,
+    DialogPlatform? dialogPlatform,
   }) =>
       showDialog(
         title: title,
