@@ -14,10 +14,19 @@ class SnackbarService {
     return Get.key;
   }
 
-  Map<dynamic, SnackbarConfig> _customSnackbarConfigs =
-      Map<dynamic, SnackbarConfig>();
+  Map<dynamic, SnackbarConfig?> _customSnackbarConfigs =
+      Map<dynamic, SnackbarConfig?>();
 
-  SnackbarConfig _snackbarConfig;
+  Map<dynamic, SnackbarConfig Function()?> _customSnackbarConfigBuilders =
+      Map<dynamic, SnackbarConfig Function()?>();
+
+  Map<dynamic, Widget Function(String?, Function?)?> _mainButtonBuilder =
+      Map<dynamic, Widget Function(String?, Function?)?>();
+
+  SnackbarConfig? _snackbarConfig;
+
+  /// Checks if there is a snackbar open
+  bool? get isOpen => Get.isSnackbarOpen;
 
   /// Saves the [config] to be used for the [showSnackbar] function
   void registerSnackbarConfig(SnackbarConfig config) =>
@@ -28,28 +37,41 @@ class SnackbarService {
       'Prefer to use the registerCustomSnackbarConfig() method. Will be removed in future release')
   void registerCustomSnackbarconfig({
     @required dynamic customData,
-    @required SnackbarConfig config,
+    @required SnackbarConfig? config,
   }) =>
       registerCustomSnackbarConfig(
         variant: customData,
         config: config,
       );
 
-  /// Saves the [config] against the value of [variant]
-  void registerCustomSnackbarConfig({
+  /// Registers a builder that will be used when showing a matching variant value. The builder
+  /// function takes in a [String] to display as the title and a `Function` to be used to the
+  /// onTap callback
+  void registerCustomMainButtonBuilder({
     @required dynamic variant,
-    @required SnackbarConfig config,
+    @required Widget Function(String?, Function?)? builder,
   }) =>
-      _customSnackbarConfigs[variant] = config;
+      _mainButtonBuilder[variant] = builder;
+
+  /// Saves the [config] against the value of [variant]. A [configBuilder] can also be
+  /// supplied which will be chosen over the config for the same variant when requested.
+  void registerCustomSnackbarConfig({
+    required dynamic variant,
+    SnackbarConfig? config,
+    SnackbarConfig Function()? configBuilder,
+  }) {
+    _customSnackbarConfigs[variant] = config;
+    _customSnackbarConfigBuilders[variant] = configBuilder;
+  }
 
   /// Shows a snack bar with the details passed in
   void showSnackbar({
-    String title,
-    @required String message,
-    Function(dynamic) onTap,
+    String title = '',
+    required String message,
+    Function(dynamic)? onTap,
     Duration duration = const Duration(seconds: 3),
-    String mainButtonTitle,
-    Function onMainButtonTapped,
+    String? mainButtonTitle,
+    void Function()? onMainButtonTapped,
   }) {
     final mainButtonWidget = _getMainButtonWidget(
       mainButtonTitle: mainButtonTitle,
@@ -64,18 +86,22 @@ class SnackbarService {
           ? Text(
               title,
               style: TextStyle(
-                  color: _snackbarConfig.titleColor,
-                  fontWeight: FontWeight.w800,
-                  fontSize: 16),
+                color: _snackbarConfig?.titleColor,
+                fontWeight: FontWeight.w800,
+                fontSize: 16,
+              ),
+              textAlign: _snackbarConfig?.titleTextAlign ?? TextAlign.left,
             )
           : null,
       messageText: _snackbarConfig?.messageColor != null
           ? Text(
               message,
               style: TextStyle(
-                  color: _snackbarConfig.messageColor,
-                  fontWeight: FontWeight.w300,
-                  fontSize: 14),
+                color: _snackbarConfig?.messageColor,
+                fontWeight: FontWeight.w300,
+                fontSize: 14,
+              ),
+              textAlign: _snackbarConfig?.messageTextAlign ?? TextAlign.left,
             )
           : null,
       colorText: _snackbarConfig?.textColor ?? Colors.white,
@@ -92,14 +118,14 @@ class SnackbarService {
     );
   }
 
-  Future showCustomSnackBar({
-    @required String message,
+  Future? showCustomSnackBar({
+    required String message,
     @deprecated dynamic customData,
     dynamic variant,
-    String title,
-    String mainButtonTitle,
-    Function onMainButtonTapped,
-    Function onTap,
+    String? title,
+    String? mainButtonTitle,
+    void Function()? onMainButtonTapped,
+    Function? onTap,
     Duration duration = const Duration(seconds: 1),
   }) {
     // TODO: Remove customData in the future release and set variant as required
@@ -109,45 +135,56 @@ class SnackbarService {
       'No variant defined, you should provide the variant property to show a custom snackbar',
     );
 
-    var snackbarConfig = _customSnackbarConfigs[snackbarVariant];
+    final snackbarConfigSupplied = _customSnackbarConfigs[snackbarVariant];
+    final snackbarConfigBuilder =
+        _customSnackbarConfigBuilders[snackbarVariant];
+
+    final snackbarConfig = snackbarConfigBuilder != null
+        ? snackbarConfigBuilder()
+        : snackbarConfigSupplied;
 
     if (snackbarConfig == null) {
       throw CustomSnackbarException(
-        'No config found for $snackbarVariant make sure you have called registerCustomConfig. See [README LINK HERE] for implementation details.',
+        'No config found for $snackbarVariant make sure you have called registerCustomConfig with a config or a builder. See [https://pub.dev/packages/stacked_services#custom-styles] for implementation details.',
       );
     }
 
-    final mainButtonWidget = _getMainButtonWidget(
-      mainButtonTitle: mainButtonTitle,
-      onMainButtonTapped: onMainButtonTapped,
-      config: snackbarConfig,
-    );
+    final mainButtonBuilder = _mainButtonBuilder[variant];
+    final hasMainButtonBuilder = mainButtonBuilder != null;
+
+    final mainButtonWidget = hasMainButtonBuilder
+        ? mainButtonBuilder!(mainButtonTitle, onMainButtonTapped)
+        : _getMainButtonWidget(
+            mainButtonTitle: mainButtonTitle,
+            onMainButtonTapped: onMainButtonTapped,
+            config: snackbarConfig,
+          );
 
     final getBar = GetBar(
-      titleText: Text(
-        title,
-        style: TextStyle(
-          color: snackbarConfig?.titleColor ??
-              snackbarConfig?.textColor ??
-              Colors.white,
-          fontWeight: FontWeight.w800,
-          fontSize: 16,
-        ),
-      ),
+      titleText: title != null
+          ? Text(
+              title,
+              style: TextStyle(
+                color: snackbarConfig.titleColor ?? snackbarConfig.textColor,
+                fontWeight: FontWeight.w800,
+                fontSize: 16,
+              ),
+              textAlign: snackbarConfig.titleTextAlign,
+            )
+          : null,
       messageText: Text(
         message,
         style: TextStyle(
-          color: snackbarConfig?.messageColor ??
-              snackbarConfig?.textColor ??
-              Colors.white,
+          color: snackbarConfig.messageColor ?? snackbarConfig.textColor,
           fontWeight: FontWeight.w300,
           fontSize: 14,
         ),
+        textAlign: snackbarConfig.messageTextAlign,
       ),
       icon: snackbarConfig.icon,
       shouldIconPulse: snackbarConfig.shouldIconPulse,
       maxWidth: snackbarConfig.maxWidth,
-      margin: snackbarConfig.margin,
+      margin: snackbarConfig.margin ?? EdgeInsets.zero,
       padding: snackbarConfig.padding,
       borderRadius: snackbarConfig.borderRadius,
       borderColor: snackbarConfig.borderColor,
@@ -157,7 +194,7 @@ class SnackbarService {
       boxShadows: snackbarConfig.boxShadows,
       backgroundGradient: snackbarConfig.backgroundGradient,
       mainButton: mainButtonWidget,
-      onTap: onTap,
+      onTap: (object) => onTap!(),
       duration: duration,
       isDismissible: snackbarConfig.isDismissible,
       dismissDirection: snackbarConfig.dismissDirection,
@@ -181,7 +218,7 @@ class SnackbarService {
       return getBar.show();
     } else {
       Completer completer = new Completer();
-      WidgetsBinding.instance.addPostFrameCallback((_) async {
+      WidgetsBinding.instance?.addPostFrameCallback((_) async {
         final result = await getBar.show();
         completer.complete(result);
       });
@@ -189,16 +226,16 @@ class SnackbarService {
     }
   }
 
-  Widget _getMainButtonWidget({
-    String mainButtonTitle,
-    Function onMainButtonTapped,
-    SnackbarConfig config,
+  TextButton? _getMainButtonWidget({
+    String? mainButtonTitle,
+    void Function()? onMainButtonTapped,
+    SnackbarConfig? config,
   }) {
     if (mainButtonTitle == null) {
       return null;
     }
 
-    return FlatButton(
+    return TextButton(
       child: Text(
         mainButtonTitle,
         style: TextStyle(
