@@ -2,15 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class SharedPreferencesService {
-  static SharedPreferencesService _instance;
+  static SharedPreferencesService? _instance;
 
   static Future<SharedPreferencesService> getInstance() async {
     if (_instance == null) {
       _instance =
           SharedPreferencesService._(await SharedPreferences.getInstance());
     }
-
-    return _instance;
+    return _instance!;
   }
 
   final SharedPreferences _preferences;
@@ -19,11 +18,11 @@ class SharedPreferencesService {
   static const _ThemeIndexKey = 'user_key';
   static const _UserThemeModeKey = 'user_theme_mode_key';
 
-  int get themeIndex => _getFromDisk(_ThemeIndexKey);
+  int? get themeIndex => _getFromDisk(_ThemeIndexKey);
 
-  set themeIndex(int value) => _saveToDisk(_ThemeIndexKey, value);
+  set themeIndex(int? value) => _saveToDisk(_ThemeIndexKey, value);
 
-  ThemeMode get userThemeMode {
+  ThemeMode? get userThemeMode {
     var userThemeString = _getFromDisk(_UserThemeModeKey);
     if (userThemeString == ThemeMode.dark.toString()) {
       return ThemeMode.dark;
@@ -36,7 +35,7 @@ class SharedPreferencesService {
     return null;
   }
 
-  set userThemeMode(ThemeMode value) {
+  set userThemeMode(ThemeMode? value) {
     if (value == null) {
       _saveToDisk(_UserThemeModeKey, value);
     } else {
