@@ -1,8 +1,14 @@
 library stacked;
 
-export 'src/base_view_models.dart';
-export 'src/index_tracking_viewmodel.dart';
-export 'src/reactive_service_mixin.dart';
-export 'src/view_model_builder.dart';
-export 'src/view_model_builder_widget.dart';
-export 'src/view_model_widget.dart';
+export 'src/state_management/base_view_models.dart';
+export 'src/state_management/index_tracking_viewmodel.dart';
+export 'src/state_management/form_viewmodel.dart';
+export 'src/state_management/reactive_service_mixin.dart';
+export 'src/state_management/view_model_builder.dart';
+export 'src/state_management/view_model_builder_widget.dart';
+export 'src/state_management/view_model_widget.dart';
+
+export 'src/code_generation/router/extended_navigator.dart';
+export 'src/code_generation/router/router_base.dart';
+export 'src/code_generation/router/route_def.dart';
+export 'src/code_generation/stacked_locator.dart';
