@@ -10,7 +10,7 @@ class SharedPreferencesServiceMock extends Mock
 class StatusBarServiceMock extends Mock implements StatusBarService {}
 
 SharedPreferencesService getAndRegisterSharedPreferencesServiceMock(
-    {int themeIndex, ThemeMode userThemeMode}) {
+    {int? themeIndex, ThemeMode? userThemeMode}) {
   _removeRegistrationIfExists<SharedPreferencesService>();
   var service = SharedPreferencesServiceMock();
 
@@ -31,8 +31,8 @@ StatusBarService getAndRegisterStatusBarServiceMock() {
 // Call this before any service registration helper. This is to ensure that if there
 // is a service registered we remove it first. We register all services to remove boiler plate from tests
 void _removeRegistrationIfExists<T>() {
-  if (locator.isRegistered<T>()) {
-    locator.unregister<T>();
+  if (locator.isRegistered()) {
+    locator.unregister();
   }
 }
 
